backend/pyproject.toml
[project]
name = "baluhost-backend"
version = "0.1.0"
description = "Mocked NAS management backend built with FastAPI"
authors = [{ name = "Baluhost" }]
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<0.31.0",
  "pydantic>=2.6.0,<3.0.0",
  "email-validator>=2.1.0,<3.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "pyjwt>=2.9.0,<3.0.0",
  "sqlalchemy>=2.0.0,<3.0.0",
  "alembic>=1.13.0,<2.0.0",
  "passlib[bcrypt]>=1.7.4,<2.0.0",
  "bcrypt>=3.2.0,<4.0.0",
  "python-multipart>=0.0.9,<0.0.10",
  "psutil>=5.9.0,<6.0.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0,<0.28.0",
  "fastapi[all]>=0.115.0,<1.0.0",
  "pytest>=7.4.0,<8.0.0",
  "pytest-asyncio>=0.23.0,<0.24.0",
  "ruff>=0.6.0,<0.7.0"
]

[project.scripts]
baluhost-backend = "app.main:run"

[build-system]
requires = ["setuptools>=67.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["storage*", "tmp*", "tests*"]
